server/voyager_server/requirements.txt
django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
psycopg2==2.9.4